as.character.readtext {readtext}	R Documentation

return only the texts from a readtext object

Description

An accessor function to return the texts from a readtext object as a character vector, with names matching the document names.

Usage

## S3 method for class 'readtext'
as.character(x, ...)

Arguments

x
the readtext object whose texts will be extracted

...
further arguments passed to or from other methods

[Package readtext version 0.71 Index]