encoding {readtext}	R Documentation

detect the encoding of texts

Description

Detect the encoding of texts in a character readtext object and report on the most likely encoding for each document. Useful in detecting the encoding of input texts, so that a source encoding can be (re)specified when inputting a set of texts using readtext, prior to constructing a corpus.

Usage

encoding(x, verbose = TRUE, ...)

Arguments

x
character vector, corpus, or readtext object whose texts' encodings will be detected.

verbose
if FALSE, do not print diagnostic report

...
additional arguments passed to stri_enc_detect

Details

Based on stri_enc_detect, which is in turn based on the ICU libraries. See the ICU User Guide, http://userguide.icu-project.org/conversion/detection.

Examples

encoding(data_char_encodedtexts)
# show detected value for each text, versus known encoding
data.frame(labelled = names(data_char_encodedtexts), 
           detected = encoding(data_char_encodedtexts)$all)

## Not run: # Russian text, Windows-1251
myreadtext <- readtext("http://www.kenbenoit.net/files/01_er_5.txt")
encoding(myreadtext)

## End(Not run)

[Package readtext version 0.71 Index]