print.readtext {readtext}	R Documentation

print method for a readtext object

Description

Print a readtext object in a nicely formatted way.

Usage

## S3 method for class 'readtext'
print(x, n = 6L, text_width = 10L, ...)

Arguments

x
the readtext object to be printed

n
a single integer, the number of rows of a readtext object to print.

text_width
number of characters to display of the text field

...
not used here

[Package readtext version 0.71 Index]