sort_fields {readtext}	R Documentation

Move text to the first column and set types to document variables

Description

Move text to the first column and set types to document variables

Usage

sort_fields(x, path, text_field, impute_types = TRUE)

Arguments

x
data.frame; contains texts and document variables

paht
character; file path from which x is created; only use in error message

text_filed
numeric or character; indicate position of a text column in x

inpute_types
logical; if TRUE, set types of variables automatically

[Package readtext version 0.71 Index]